+++ packages/narrat/src/stores/screens-store.ts
@@ -1,3 +1,4 @@
+import { ScreensConfig } from '@/config/screens-config';
 import { deepCopy } from '@/utils/data-helpers';
 import { error, warning } from '@/utils/error-handling';
 import {
@@ -7,7 +8,6 @@ import {
 } from '@/utils/transition';
 import deepmerge from 'deepmerge';
 import { defineStore } from 'pinia';
-import { Config } from '../config';
 
 export type ButtonStateValue = boolean | 'hidden' | 'greyed';
 export interface ButtonsState {
@@ -32,12 +32,6 @@ export type ScreenSave = {
   buttons: ButtonsState;
 };
 
-// Create a pinia store named screens with a state using the type ScreenState, with actions:
-// setScreen(screen: string): Sets the current screen to the given screen
-// setButtons(buttons: { [key: string]: ButtonConfig }): Adds buttons to the buttons state by using the values in the buttons config object
-// changeButton(button: string, newValue: boolean): Changes the value of a button in the buttons state
-// generateSaveData(): Function that generates a ScreenState object from the data in the state
-// loadSaveData(data: ScreenState): Function that loads the data into the state using a deepmerge of current state value and new value
 export const useScreens = defineStore('screens', {
   state: () =>
     ({
@@ -115,28 +109,23 @@ export const useScreens = defineStore('screens', {
         }
       });
     },
-    setButtons(config: Config) {
-      const { buttons: buttonsConfig, screens: screensConfig, images } = config;
+    setButtons(config: ScreensConfig) {
+      const { buttons, screens } = config;
 
-      for (const key in buttonsConfig) {
+      for (const key in buttons) {
         this.buttons[key] = {
-          state: buttonsConfig[key].enabled,
+          state: buttons[key].enabled,
         };
       }
       // Support for inline button conf in screens as it's easier for end users
       // We basically copy the button's config to the buttons config object if we find any buttons defined inline in a screen
-      for (const key in screensConfig) {
-        if (screensConfig[key].buttons) {
-          const screen = screensConfig[key];
-          if (screen.background && images[screen.background]) {
-            // Also defaulting to specifying screen backgrounds inline now
-            screen.background = images[screen.background];
-          }
-          for (const index in screen.buttons) {
-            const button = screen.buttons[index];
+      for (const key in screens) {
+        const screen = screens[key];
+        if (screen.buttons) {
+          for (const [index, button] of screen.buttons.entries()) {
             // If the button is a config object, add it to the buttons config, and also create its state
             if (typeof button === 'object') {
-              buttonsConfig[button.id] = button;
+              config.buttons[button.id] = button;
               // Change the inline config to be a string again
               screen.buttons[index] = button.id;
               this.buttons[button.id] = {